requirements.txt
pillow==11.0.0
pyyaml==6.0.2
comfyui_api @ git+https://github.com/DeInfernal/comfyui_api.git@v1.0.1
